pyproject.toml
[tool.poetry]
name = "sqlalchemy-cassandra"
version = "0.1.0"
description = "SQLAlchemy dialect for Apache Cassandra"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "sqlalchemy_cassandra", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.0"
cassandra-driver = "^3.28.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[tool.poetry.plugins."sqlalchemy.dialects"]
cassandra = "sqlalchemy_cassandra.dialect:CassandraDialect"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 